scripts/pyproject.toml
[project]
name = "hoa-ci"
version = "0.1.0"
description = "Build course pages from HOA."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pygithub>=2.8.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]
